Issue-159 - Move connect_to_router to its own file
