+++ extension/src/main.rs
@@ -17,6 +17,7 @@ use crate::prelude::*;
 mod app_request;
 mod app_start;
 mod cert;
+mod connect_to_router;
 mod counter_drop;
 mod endpoint;
 mod lambda_request;
